JBTM-2224 Move Compensations dependency adder to transactions subsytem
