+++ feature-pack/src/main/resources/modules/system/layers/base/org/jboss/as/transactions/main/module.xml
@@ -55,5 +55,6 @@
         <module name="org.jboss.as.jacorb"/>
         <module name="org.jboss.jandex"/>
         <module name="org.jboss.modules"/>
+        <module name="org.jboss.narayana.compensations"/>
     </dependencies>
 </module>
